ml/requirements.txt
joblib==1.3.2
pandas==1.5.3
scikit-learn==1.4.0
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
fmt==0.3.1
xgboost==2.0.3
category_encoders==2.6.3